+++ src/org/ods/build_service/GitService.groovy
@@ -1,5 +1,4 @@
-package org.ods.service
-
+package org.ods.build_service
 
 class GitService {
